Time: 2026 ms (5.04%), Space: 99.2 MB (16.98%) - LeetHub

+++ 3066-minimum-operations-to-exceed-threshold-value-ii/3066-minimum-operations-to-exceed-threshold-value-ii.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int minOperations(vector<int>& nums, int k) {
-        priority_queue<int,vector<int>,greater<int>>temp;
+        priority_queue<long long,vector<long long>,greater<long long>>temp;
         for(int i=0;i<nums.size();i++)
         {
             temp.push(nums[i]);
@@ -11,11 +11,11 @@ public:
         {
             if(temp.top()<k)
             {
-                int s1=temp.top();
+                long long s1=temp.top();
                 temp.pop();
-                int s2=temp.top();
+                long long s2=temp.top();
                 temp.pop();
-                int res=s1*2+s2;
+                long long res=s1*2+s2;
                 temp.push(res);
                 // temp.push(res);
                 cout<<s1<<" "<<s2<<" "<<res<<endl;
